SNOW-1794920: Test more Series.map cases and refactor Series apply and map tests.
